Separate slurm and slurmdb REST model generation
